% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_selectivity.R
\name{add_selectivity}
\alias{add_selectivity}
\title{Add gillnet selectivity to a data frame}
\usage{
add_selectivity(
  df,
  species_col = "species_code",
  length_col = "length_mm",
  design_col = "sample_design_code",
  sample_event_col = "Sample_event",
  method_col = "method",
  out_col = "select",
  default_design = "SGN7"
)
}
\arguments{
\item{df}{A data frame containing fish records.}

\item{species_col}{Name of the species column.}

\item{length_col}{Name of the fork length column (mm).}

\item{design_col}{Name of the sample/net design column.}

\item{sample_event_col}{Name of Sample_event column (optional).}

\item{method_col}{Name of the capture method column.}

\item{out_col}{Name of the output column to add.}

\item{default_design}{Design code used when design is missing.}
}
\description{
Adds a selectivity column to a data frame using \code{\link{GN_select}}.
}
\details{
Rules:
\itemize{
  \item If \code{method_col} is not \code{"GN"}, selectivity is \code{NA}.
  \item If \code{Sample_event} is present and non-missing, event-level
        selectivity is used.
  \item Otherwise, \code{sample_design_code} is used.
  \item If GN and design is missing/NA, \code{default_design} is used.
}
}
