% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{approx_select_spp}
\alias{approx_select_spp}
\title{Approximate species mapping for gillnet selectivity}
\format{
A data frame with the following columns:
\describe{
  \item{species_code}{Character. Species code observed in biological samples.}
  \item{select_spp}{Character. Species code of the selectivity model to use
    when applying gillnet selectivity.}
  \item{match_level}{Character. Description of how the proxy species was chosen
    (e.g., taxonomic match level).}
}
}
\usage{
data(approx_select_spp)
}
\description{
A lookup table mapping observed fish \code{species_code} values to the
selectivity model species actually used when applying gillnet selectivity.
}
\details{
This table is used when a species does not have a fitted selectivity curve
in \code{\link{sel_lookup}}. In such cases, selectivity is approximated using
a proxy species (\code{select_spp}) that *does* have a curve.

The table is generated internally and stored in the package \code{data/}
directory. It is not intended to be edited manually.


\code{approx_select_spp} is used by \code{\link{GN_select}} to map species to
an available selectivity curve in \code{\link{sel_lookup}}.
}
\seealso{
\code{\link{sel_lookup}},
\code{\link{GN_select}}
}
\keyword{internal}
