% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cam_xdata.R
\name{Cam_xdata}
\alias{Cam_xdata}
\title{Cam_xdata}
\usage{
Cam_xdata(data = Edata)
}
\arguments{
\item{data}{a data set to search for ground or air based verification counts of camera data. defaults to data, as this function is run within the Effort2R() function}
}
\description{
A function to create data subsets to estimate expansion factors for cameras where there has been ground counts to verify. Effort estimates from Camera data is compared to estimates from AIR/GRD with EffortEst function and the ratio of those is considered the expansion.
}
\examples{

DP2R::Cam_xdata()

}
\keyword{DataPond}
\keyword{SPDT;}
