% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEANreleases.R
\name{CLEANreleases}
\alias{CLEANreleases}
\title{CLEANreleases}
\usage{
CLEANreleases()
}
\description{
A function to load, clean and standardize release data for analysis.
Function only usable with connection to DataPond.
}
\details{
This is a non-exported function to simplify code using the releases table to understand all possible stocked groups that could be in a lake at any given time.
In cases where clips are unique, then fields for age, strain, genotype are updated.
Previous version of this function was within SPDT as SPDTreleases
Ultimately, as upload filters and cleaning are improved in the main database, this function could become obsolete.
}
\examples{
#' Must be connected to VPN if working remotely

CLEANreleases()
}
\keyword{DP2R;}
\keyword{clips}
\keyword{releases;}
\keyword{stocked;}
