% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DayTypes}
\alias{DayTypes}
\title{DayTypes Dataset}
\format{
A data frame with 3 columns:
\describe{
  \item{Date}{Date in "YYYY-MM-DD" format.}
  \item{DayType}{A factor with levels: "Weekday", "Weekend", "Holiday".}
  \item{Description}{A short description of the day type.}
}
}
\source{
Generated internally for the DP2R package.
}
\usage{
data(DayTypes)
}
\description{
A dataset indicating different types of days used in the effort analysis.
}
\details{
This dataset categorizes days as "Weekday", "Weekend", or "Holiday", which are
used for modeling effort variations in the DP2R package.
}
\examples{
data(DayTypes)
head(DayTypes)

}
