% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP2R.R
\name{DP2R}
\alias{DP2R}
\title{DP2R}
\usage{
DP2R(
  Tables = c("vwPaAssessEvent", "vwLegacyRelease", "vwIndividualFish", "vwCollectCount",
    "vwFishCollection", "vwWaterbody", "Species", "SampleDesign_MeshSizeCode"),
  exclude_types = c("geography", "varbinary", "ntext"),
  envir = .GlobalEnv,
  as.raw = FALSE
)
}
\arguments{
\item{Tables}{A text vector of DataPond tables names to be pulled into the R environment. If unsure of the names of all tables, establish a DataPond connection (conn) and then run: 'DBI::dbListTables(conn)' or review connections tab in RStudio. The default tables are the 'views' (preceded by vw), which would typically be used in analyses of gillnet assessment data.}

\item{exclude_types}{A text string or vector describing data types to omit from the data set. The default is "geography" because R does not handle that datatype. That means any columns with point data in the geography datatype are removed from the dataset (there is still lat-long data that can be used).}

\item{envir}{Environment to assign results into (default: .GlobalEnv).}

\item{as.raw}{Logical; if TRUE, keep data in raw DB format (only excluded-type filtering is applied; no renaming/coalescing/aggregation). Default: FALSE.}

\item{DB}{A text string of "DataPond" (default production data) or "DataPond_STAGE" which is staged data to be reviewed before acceptance into production.}
}
\description{
A function to bring data into R from DataPond for analysis.
It is important that you establish a personal password protected connection (and assign as 'conn' object before running this function. Use example code below and/or you may have it set up under Connections tab in RStudio after running it once.See: https://github.com/r-dbi/odbc/blob/main/README.md
Function only usable by FFSBC staff who have a direct or vpn connection to SLD.Replaces old 'SLD2R()' function for the old database.
}
\examples{
  conn <- DBI::dbConnect(drv = odbc::odbc(),
  Driver = 'SQL Server',
  server = 'tcp:gofishbc.database.windows.net,1433',
  database = 'DataPond',
  uid = uid,
  pwd = pwd)

DP2R()

}
\keyword{DataPond}
\keyword{SPDT;}
