% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DP2R_Effort_Model}
\alias{DP2R_Effort_Model}
\title{DP2R effort model (internal file)}
\description{
A generalized linear mixed-effects model (GLMM) used to estimate fishing
effort based on day type and other covariates.
}
\details{
This model is stored as a compressed \code{.qs2} file in the package
\code{data/} directory and is not loaded automatically. It is only required
when updating or refitting effort estimates.


To load the model manually:
\preformatted{
  fit <- qs2::qs_read(system.file("data", "DP2R_Effort_Model.qs2", package = "DP2R"))
}
}
\seealso{
\code{\link{EffortEst}}
}
\keyword{internal}
