% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DP2R_Ice_Model}
\alias{DP2R_Ice_Model}
\title{DP2R ice fishing effort model (internal file)}
\description{
A generalized linear mixed-effects model (GLMM) used to estimate ice fishing
effort.
}
\details{
This model is stored as a compressed \code{.qs2} file in the package
\code{data/} directory and is not loaded automatically.


To load the model manually:
\preformatted{
  fit <- qs2::qs_read(system.file("data", "DP2R_Ice_Model.qs2", package = "DP2R"))
}
}
\seealso{
\code{\link{EffortEst}}
}
\keyword{internal}
