% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effort2R.R
\name{Effort2R}
\alias{Effort2R}
\title{Effort2R}
\usage{
Effort2R()
}
\description{
A function to bring Effort data into R from DataPond for analysis.
It is important that you establish a personal password protected connection (and assign as 'conn' object before running this function. Use example code below and/or you may have it set up under Connections tab in RStudio after running it once.See: https://github.com/r-dbi/odbc/blob/main/README.md
Function only usable by FFSBC staff who have a direct or vpn connection to DataPond
}
\examples{
  conn <- DBI::dbConnect(drv = odbc::odbc(),
  Driver = 'SQL Server',
  server = 'tcp:gofishbc.database.windows.net,1433',
  database = 'DataPond',
  uid = uid,
  pwd = pwd)

DP2R::Effort2R()

}
\keyword{DataPond}
\keyword{SPDT;}
