% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffortClean.R
\name{EffortClean}
\alias{EffortClean}
\title{EffortClean}
\usage{
EffortClean()
}
\description{
A function to bring Effort data into R from DataPond for analysis.
A temporary function to clean the Effort data which has duplicates and generally useless data.
}
\keyword{DataPond}
\keyword{SPDT;}
