% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac.data.R
\name{fac.data}
\alias{fac.data}
\title{fac.data}
\usage{
fac.data(data, varlist = list())
}
\arguments{
\item{data}{a data.frame object}

\item{varlist}{a list of variable names and their reference level to factor in the format "name" = "ref level". If no reference level is required then "name" = NULL must be specficied.}
}
\description{
A function analyze Effort counts and estimate total annual effort in angler days.
This function builds on other DP2R functions to gather data and fit an effort model to observed data. THe final step is to predict effort in all the unobserved strata withthe effort model.
}
\examples{

DP2R::DP2R(Tables = c("vwEffort"))
vwEffort = DP2R::fac.data(vwEffort, vars = list("year" = NULL, "month" = "5", "hour" = "12", "daytype" = "WE", "lakeview_yr" = NULL, "lake_hr" = NULL))

}
\keyword{DP2R;}
\keyword{DataPond}
