% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit}
\alias{fit}
\title{DP2R_Effort_Model}
\format{
An object of class `glmerMod`, loaded into the environment as `fit`.
}
\source{
Generated by fitting a GLMM using the `lme4` package.
}
\usage{
data("DP2R_Effort_Model", package = "DP2R")
       load("data/DP2R_Effort_Model.rda")
}
\description{
A `glmer` model object for predicting effort based on day types and other variables.
}
\details{
This object contains a generalized linear mixed-effects model (GLMM) fitted using the
`lme4::glmer()` function. It is stored in the package but is not exported. You can
load this object into your environment using the `data()` function or the `load()` function.
}
\examples{
# To load the model, you can use:
data("DP2R_Effort_Model", package = "DP2R")

# After loading, you can use the model as follows:
summary(fit)
}
\keyword{datasets}
