% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_wt_cf.R
\name{fl_wt_cf}
\alias{fl_wt_cf}
\title{Calculate condition factor, length, or weight}
\usage{
fl_wt_cf(fl_mm = NULL, wt_g = NULL, cf = NULL)
}
\arguments{
\item{fl_mm}{Numeric. Fish fork-length in millimetres.}

\item{wt_g}{Numeric. Fish weight in grams.}

\item{cf}{Numeric. Condition factor.}
}
\value{
Numeric value of the missing parameter.
}
\description{
CF = (W / L^3) * 100000   where W = weight in grams, L = fork length in mm
}
\details{
Provide exactly TWO of: fl_mm, wt_g, cf.
The function returns the missing third value.
}
