% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GN_select.R
\name{GN_select}
\alias{GN_select}
\title{Gillnet selectivity at fork length}
\usage{
GN_select(species, length_mm, sample_design_code = "SGN7", Sample_event = NULL)
}
\arguments{
\item{species}{Species code(s).}

\item{length_mm}{Fork length(s) in mm.}

\item{sample_design_code}{Net/sample design code(s). Defaults to \code{"SGN7"}.}

\item{Sample_event}{Optional Sample_event id(s). If provided, curves are pulled from
\code{sel_lookup_event} instead of \code{sel_lookup}.}
}
\value{
Numeric vector of selectivity values (0--1) or NA.
}
\description{
Returns relative gillnet selectivity (0--1) for one or more fish given
species code, fork length (mm), and either sample design code OR
Sample_event (which can have multiple sample_design_codes in the event).
}
\details{
Curves are stored on a shared length grid \code{sel_classes}.
Species without direct curves are mapped to proxy species using \code{approx_select_spp}.
}
\examples{
GN_select("RB", 312, sample_design_code = "SGN7")
GN_select("RB", 312, Sample_event = "GN_00394CHIL_SON")

}
