% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakesum}
\alias{lakesum}
\title{# ----------------------------------------------------------------------
lakesum Dataset}
\format{
A data frame with 10 columns:
\describe{
  \item{region}{managment region, character}
  \item{WBID}{Waterbody Identifie, character}
  \item{gazetted_name}{Gazetted waterbody name}
  \item{N_years}{the number of different calendar years with effort estimates, integer}
  \item{min_year}{the earliest year with effort estimates, integer}
  \item{max_year}{the nmost recent year effort estimates, integer}
  \item{Methods}{A character string describing the different methods used to count effort CAM, AIR, GRD}
  \item{mean_AD}{The average estimated total angler days across years assessed, numeric}
  \item{marker_size}{A value equal to mean_AD for plotting}
  \item{AD_percentile}{The percentile rank, province wide, of the average anglers days estiamted for a given lake.}
  \item{lake_latitude}{coordinate of lake, numeric}
  \item{lake_longitude}{coordinate of lake, numeric}
}
}
\source{
Generated internally for the DP2R package.
}
\usage{
data(lakesum)
}
\description{
A data frame summarizing effort estimates for lakes over all years assessed, which underlies the Shiny application used by provincial biologists.
}
\examples{
data(lakesum)
head(lakesum)
}
\keyword{datasets}
