% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mgt_Region_Unit}
\alias{Mgt_Region_Unit}
\title{Mgt_Region_Unit Dataset}
\format{
A data frame with 10 columns:
\describe{
  \item{WATERBODY_POLY_ID}{Integer. Unique identifier for the polygon describing the lake.}
  \item{gazetted_name}{Character. Official gazetted waterbody name, if available.}
  \item{alias}{Character. Alternative or informal lake name, if available.}
  \item{WBID}{Character. Waterbody identifier.}
  \item{Management_Unit}{Character. Wildlife management unit.}
  \item{region_code}{Character. Management region code.}
  \item{RegionName}{Character. Management region full name.}
  \item{OBJECTID}{Integer. Unique identifier in provincial database.}
  \item{locale_name}{Character. Local or commonly used lake name, if available.}
  \item{neighbour_mus}{Character. Comma-separated list of adjacent WMUs.}
}
}
\source{
Generated from bcdata R package for the DP2R package and code in data-raw folder.
}
\usage{
data(Mgt_Region_Unit)
}
\description{
A data frame of regions and wildlife management units associated with lakes.
Data obtained from bcdata R package and code in raw-data folder.
}
\keyword{datasets}
