% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pred_vars}
\alias{Pred_vars}
\title{Pred_vars Dataset}
\format{
A data frame with 6 columns:
\describe{
  \item{year}{year in numeric format.}
  \item{month}{month in numeric format.}
  \item{daytype}{A factor with levels: "WD", "WE", "LWE".}
  \item{ndays}{An integer describing the number of days for that daytype category in the given month}
  \item{hour}{hour in character format}
  \item{weather_code}{A character string describing the weather catgory as POOR, FAIR or GOOD.}
}
}
\source{
Generated internally for the DP2R package.
}
\usage{
data(Pred_vars)
}
\description{
A data frame summarizing frequency of variable categories for efficient model predictions used in the effort analysis.
}
\details{
This dataset categorizes daytypes, weather, hour of unobserved counts to extrapolate effort estimates to a full season.Used within EffortEst function.
}
\examples{
data(Pred_vars)
head(Pred_vars)
}
\keyword{datasets}
