% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Releases2R.R
\name{Releases2R}
\alias{Releases2R}
\title{Releases2R}
\usage{
Releases2R()
}
\description{
Ultimately, as upload filters and cleaning are improved in the main database, this function will become obsolete.
}
\examples{
Releases2R()
}
\keyword{DP2R}
\keyword{SPDT;}
