% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sel_classes}
\alias{sel_classes}
\title{Fork-length grid for gillnet selectivity curves}
\format{
A numeric vector of fork lengths (mm).
}
\source{
Defined internally for the DP2R package when building \code{sel_lookup}.
}
\usage{
data(sel_classes)
}
\description{
A numeric vector defining the fork-length (mm) grid used for all selectivity
curves stored in \code{\link{sel_lookup}}.
}
\details{
All curves in \code{sel_lookup$curve} are aligned to this shared grid.
This enables fast lookup of selectivity values without interpolation.
}
\examples{
data(sel_classes)
head(sel_classes)

}
