% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTdata.R
\name{SPDTdata}
\alias{SPDTdata}
\title{SPDTdata}
\usage{
SPDTdata(
  Spp = NULL,
  Contrast = c("species_code", "Strain_rel", "SAR_cat", "Geno_rel"),
  drop_controls = NULL,
  Strains = NULL,
  Genotypes = NULL,
  filters = NULL,
  Data_source = TRUE
)
}
\arguments{
\item{Spp}{an optional character string or character vector for BC species code (e.g. "RB" or c("KO", "EB"), etc.). This will filter data to only that species.}

\item{Contrast}{a required character string describing the experimental contrast, which must be a field in the SPDTdata (e.g. "species_code", "Strain_rel", "SAR_cat", "Geno_rel" are the 3 possibilities now).
Entering a value for contrast will filter to lake years that had fish present from a co-stocking event of groups varying in your contrast variable.}

\item{drop_controls}{an optional character string or vector to drop covariates from the potenital Contrast list that would be used as controls when filtering and grouping data by the selected Contrast. Default is the full list of potential contrasts are usd as controls (i.e. none are dropped). Using all controls may be restrictive for some contrasts (eg. Comparing species it would be difficult to control for size-at-release SAR_cat because they are released at different stage-sizes)}

\item{Strains}{an optional character string or character vector describing the strain code (SPDTdata format e.g. "RB" for Rainbow Trout) for source population. This will filter to only those strains listed}

\item{Genotypes}{an optional character string or character vector to filter data to specific genotypes (e.g. 2n or AF3n)}

\item{filters}{a vector of lake-years returned from the SPDTfilter() function. SPDTfilter() allows for filtering to various non-biological aspects to the data, lakes, years, regions, et.c See?SPDTfilter()}

\item{Data_source}{a TRUE FALSE value to indicate whether to load data form the SLD, or just use data tables in the Environment.}
}
\description{
A function transferred from SPDT package to filter and format DataPond data to cases with clipped experimental stocking groups.
Function only usable by FFSBC staff who have a direct connection to DataPond.
}
\details{
This function returns the data tables from linkClips, plus new refined data sets.
"idf" is the individual level data frame, for all clipped fish meeting your filtering selections.
"gdf" is the grouped data frame, for all clipped groups meeting your filtering selections.
Our database will not likely ever have filtered summaries to this level, so this function will be used and updated in to the future.

Several new columns are present. Anything prefixed by "NetX" means that variable has been expanded to account for gillnet selectivity.
SAR if for size-at-release in grams, and SAR_cat has categorized those release sizes.
The code creating categories for the SAR_cat has bigger categories as the size increases. See ?SAR_cat
}
\examples{
#Must be connected to VPN if working remotely

#Download all data with clipped fish
SPDTdata()

#Download only KO data
SPDTdata(Spp = "KO")

#Download any data from RB, of either HF, CL or BW strain that were stocked as a strain comparison
SPDTdata(Spp = "RB", Strains = c("HF", "CL", "BW"), Contrast = "Strain")

#Download all data from lake years that had a size-at-release comparison for RB
SPDTdata(Spp = "RB", Contrast = "SAR_cat")

}
\keyword{SPDT}
