% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTfilter.R
\name{SPDTfilter}
\alias{SPDTfilter}
\title{SPDTfilter}
\usage{
SPDTfilter(
  Sample_events = NULL,
  Lk_yrs = NULL,
  Lake_names = NULL,
  WBIDs = NULL,
  Years = NULL,
  Regions = NULL,
  minN = NULL
)
}
\arguments{
\item{Sample_events}{an optional character string or character vector describing the sampling events to include in data.}

\item{Lk_yrs}{an optional character string or character vector of the Lake-years to filter the data set to. Must be in format WBID_YYYY (e.g. "01100OKAN_2020" or for multiples c("01100OKAN_2020", "01598LNTH_2018"))}

\item{Lake_names}{an optional character string or character vector describing the lakes to include in data.}

\item{WBIDs}{an optional character string or character vector describing the WBIDs to include in data.}

\item{Years}{an optional integer or vector describing the sampling years to include in data.}

\item{Regions}{an optional character string or character vector describing the management regions to include in data.}

\item{minN}{an optional integer or vector describing the minimum acceptable catch within a sample_event for the data to be included}
}
\description{
A function to facilitate rapid filtering SPDTdata to specific cases beyond the biological references (lakes, years, project)
}
\details{
This is convenience data filtering and cleaning process when doing exploratory analysis or plotting with SPDTdata.
}
\examples{
#Must be connected to VPN if working remotely

#Download all data with clipped fish
SPDTdata(Contrast = "Strain")

#Typically we would call this within the SPDTdata (or SPDTplot) call
SPDTdata(Contrast = "Strain", filters = SPDTfilter(Lake_names = "MATHEW"))

}
\keyword{SPDT}
