% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species}
\alias{Species}
\title{# ----------------------------------------------------------------------
Species Dataset}
\format{
A data frame with 23 columns:
}
\source{
DataPond.
}
\usage{
data(Species)
}
\description{
A data frame summarizing basic taxonomy and characteristics of fish species found in lakes of BC.
}
\examples{
data(Species)
head(Species)
}
\keyword{datasets}
