% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_selectivity.R
\name{add_selectivity}
\alias{add_selectivity}
\title{Add gillnet selectivity to a data frame}
\usage{
add_selectivity(
  df,
  species_col = "species_code",
  length_col = "length_mm",
  design_col = "sample_design_code",
  method_col = "method",
  out_col = "select",
  default_design = "SGN7"
)
}
\arguments{
\item{df}{A data frame containing fish records.}

\item{species_col}{Name of the species column. Defaults to \code{"species_code"}.}

\item{length_col}{Name of the fork length column (mm). Defaults to \code{"length_mm"}.}

\item{design_col}{Name of the sample/net design column. Defaults to \code{"sample_design_code"}.}

\item{method_col}{Name of the capture method column. Defaults to \code{"method"}.}

\item{out_col}{Name of the output column to add. Defaults to \code{"select"}.}

\item{default_design}{Design code used when \code{design_col} is missing/NA
for \code{method == "GN"}. Defaults to \code{"SGN7"}.}
}
\value{
\code{df} with an added column \code{out_col}.
}
\description{
Adds a selectivity column to a data frame using \code{\link{GN_select}}.
}
\details{
Rules:
\itemize{
  \item If \code{method_col} is not \code{"GN"}, selectivity is \code{NA}.
  \item If \code{method_col} is \code{"GN"} and \code{design_col} is missing/NA,
        \code{default_design} is used.
}
}
\examples{
idf <- add_selectivity(idf, out_col = "NetX")


}
