% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effortmodel.R
\name{Effortmodel}
\alias{Effortmodel}
\title{Effortmodel}
\usage{
Effortmodel(
  data = NULL,
  update_or_fit = "update",
  model_path = "data/DP2R_Effort_Model.qs2",
  add_camX = TRUE
)
}
\arguments{
\item{data}{a data.frame object of raw counts that wil be used to convert to annual effort estimates. Default value is Edata, which is generated form the DP2R functions Effort2R().}

\item{update_or_fit}{a character string "update" or "fit" to indicate whether to update the current effort model with new data (i.e. model fitting process builds on old model by fitting new data points and takes about 5min), or (fit") completely refit the model form scratch (takes about 20min to run.)}

\item{model_path}{path to current effort model to use for predictions of unobserved time strata. Two models exist, one for open water season: "data/DP2R_Effort_Model.qs2", and one for ice fishing "data/DP2R_Ice_Model.qs2"}
}
\description{
A function analyze Effort counts and estimate total annual effort in angler days.
It is important that you establish a personal password protected connection (and assign as 'conn' object before running this function. Use example code below and/or you may have it set up under Connections tab in RStudio after running it once.See: https://github.com/r-dbi/odbc/blob/main/README.md
Function only usable by FFSBC staff who have a direct or vpn connection to DataPond
}
\examples{

DP2R::Effort2R()
DP2R::Effortmodel()

}
\keyword{DataPond}
\keyword{SPDT;}
