% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac.data.R
\name{fac.data}
\alias{fac.data}
\title{fac.data}
\usage{
fac.data(data, varlist = list())
}
\arguments{
\item{data}{a data.frame object}

\item{varlist}{a list of variable names and their reference level to factor in the format "name" = "ref level". If no reference level is required then "name" = NULL.}
}
\description{
A simple function to take a list of variables within a data set and convert to factors and optionally specified reference levels. This is useful when setting up data for anlaysis of categorical variables in glm or glmers for example.
}
\examples{

DP2R::DP2R(Tables = c("vwEffort"))
vwEffort = DP2R::fac.data(vwEffort, vars = list("year" = NULL, "month" = "5", "hour" = "12", "daytype" = "WE", "lakeview_yr" = NULL, "lake_hr" = NULL))

}
\keyword{DP2R;}
\keyword{DataPond}
