% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lake_names}
\alias{lake_names}
\title{Lake names Dataset}
\format{
A data frame with 5 columns:
}
\source{
DataPond.
}
\usage{
data(lake_names)
}
\description{
A data frame summarizing lake names by using gazetted name where available and alias if not.Stirings ofr " Lake" or "Creek" are dropped.
Also has concatenated lake_name_WBID for quick reference.
}
\examples{
data(lake_names)
head(lake_names)

}
