% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mgt_Region_Unit}
\alias{Mgt_Region_Unit}
\title{Mgt_Region_Unit Dataset}
\format{
A data frame with 10 columns:
\describe{
  \item{WATERBODY_POLY_ID}{unique identifier for the polygon describiing the lake, integer}
  \item{GNIS_NAME_1}{Gazetted waterbody name, character}
  \item{GNIS_NAME_2}{Alias, character}
  \item{GNIS_NAME_3}{Alias, character}
  \item{WBID}{Waterbody Identifie, character}
  \item{Management_Unit}{wildlife management unit, character}
  \item{region}{managment region code, character}
  \item{RegionName}{managment region full name, character}
  \item{OBJECTID}{unique identifier provinical database, integer}
  \item{neighbour_mus}{comma separated list of wildlife managment units that are adjacent to wmu in column 5, character}
}
}
\source{
Generated from bcdata R package for the DP2R package and code in data-raw folder.
}
\usage{
data(Mgt_Region_Unit)
}
\description{
A data frame regions and wildlife management units associated with lakes.
Data obtained from bcdata R package and code in raw-data folder.
Initial function of this data was to be able to match unique lakes by name and wmu to DFO survey data.
}
\examples{
data(Mgt_Region_Unit)
head(Mgt_Region_Unit)

}
