% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Millar.R
\name{predict_Millar}
\alias{predict_Millar}
\title{Predict selectivity using Millar models by species}
\usage{
predict_Millar(
  species,
  classes = 75:750,
  rtype = NULL,
  meshSizes_in = NULL,
  theta = NULL,
  rel.power = NULL,
  theta_min_mesh = 1,
  best_models = NULL
)
}
\arguments{
\item{species}{Species code (e.g. "RB", "KO").}

\item{classes}{Vector of length classes (mm) to predict for. Default 75:650.}

\item{rtype}{Optional: Millar curve type; if NULL, taken from best_models.}

\item{meshSizes_in}{Optional: mesh sizes (mm); if NULL, taken from best_models.}

\item{theta}{Optional: parameter vector; if NULL, taken from best_models.}

\item{rel.power}{Optional: relative fishing power per mesh; default = 1 for all meshes.}

\item{theta_min_mesh}{This is a fixed parameter of the smallest mesh size used in fitting the original selectivity functions (1" mesh for nets fit in best_models), passed to rtypes_Millar(). This should not be changed for predicting nets without 1" mesh, only change if selectivity model is updated based on even smaller meshes.}

\item{best_models}{Optional: a data.frame like the DP2R best_models data frame object. If NULL, function will try to load it from package data.'best_models' must have columns: species_code, rtype, theta, meshSizes}
}
\description{
Predict selectivity using Millar models by species
}
