% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shinydata}
\alias{shinydata}
\title{DP2R_Effort_Model}
\format{
An object of class `glmerMod`, loaded into the environment as `fit`.

An object of class `glmerMod`, loaded into the environment as `fit`.

A data frame with 18 columns:
\describe{
  \item{region}{managment region, character}
  \item{WBID}{Waterbody Identifie, character}
  \item{gazetted_name}{Gazetted waterbody name}
  \item{view_location_name}{A unique name given to a specific viewpoint of a lake by camera or ground}
  \item{year}{hour in character format}
  \item{method}{A character string describing the method used to count effort CAM, AIR, GRD}
  \item{N}{The number of observations for the specified year/fishing season a lake has been assessed, integer}
  \item{spv_obs}{The total number of single person vessels observed angling in year, integer}
  \item{boats_obs}{The total number of angling boats observed in year, integer}
  \item{shore_obs}{The total number of shore anglers observed in year, integer}
  \item{spv_AD}{The estimated total angler days from single person vessels in fishing season, integer}
  \item{boats_AD}{The estimated total angler days from angling boats in fishing season, integer}
  \item{shore_AD}{The estimated total angler days from shore anglers in fishing season, integer}
  \item{Angler_days}{The estimated total angler days in fishing season, integer}
  \item{Angler_days_p_ha}{The estimated total angler days per hectare of lake area in fishing season, numeric}
  \item{area_ha}{The lake surface area in hectares, numeric}
  \item{Cam_Exp}{The expansion factor used to expand the estimated effort seen from the view_locaiton_name to the entire lake, numeric}
  \item{Exp_N}{The numerb of verification counts used to estimate the Cam_Exp, numeric}

}
}
\source{
Generated by fitting a GLMM using the `lme4` package.

Generated by fitting a GLMM using the `lme4` package.

Generated internally for the DP2R package.
}
\usage{
qs2::qs_read("data/DP2R_Effort_Model.qs2")

qs2::qs_read("data/DP2R_Ice_Model.qs2")

data(shinydata)
}
\description{
A `glmer` model object for predicting effort based on day types and other variables.
}
\details{
This object contains a generalized linear mixed-effects model (GLMM) fitted using the
`lme4::glmer()` function. It is stored in the package as a compressed .qs2 file, but is not exported. You can load this object into your environment using the `qs2::qs_read()` function.
}
\examples{
# To see model summary:
qs2::qs_read("data/DP2R_Effort_Model.qs2")

# To load the model and use the mode, assign to an object:
fit <- qs2::qs_read("data/DP2R_Effort_Model.qs2")
DP2R_Ice_Model

A `glmer` model object for predicting effort based on day types and other variables.

This object contains a generalized linear mixed-effects model (GLMM) fitted using the
`lme4::glmer()` function. It is stored in the package as a compressed .qs2 file, but is not exported. You can load this object into your environment using the `qs2::qs_read()` function.

# To see a model summary:
qs2::qs_read("data/DP2R_Ice_Model.qs2")

# To load the model and use the mode, assign to an object:
fit <- qs2::qs_read("data/DP2R_Ice_Model.qs2")
# ----------------------------------------------------------------------
shinydata Dataset

A data frame summarizing effort estimates for lakes for each year/fishing season assessed, which underlies the Shiny application used by provincial biologists. Open water fishing season is standardized from May 1st to October 31st.


data(shinydata)
head(shinydata)
}
\keyword{datasets}
